refactor: Move api settings from massa models
close #2419
