+++ massa-api/src/lib.rs
@@ -10,9 +10,8 @@ use jsonrpc_derive::rpc;
 use jsonrpc_http_server::{CloseHandle, ServerBuilder};
 use massa_consensus_exports::{ConsensusCommandSender, ConsensusConfig};
 use massa_execution_exports::ExecutionController;
-use massa_models::api::{
-    APISettings, AddressInfo, BlockInfo, BlockSummary, EndorsementInfo, EventFilter, NodeStatus,
-    OperationInfo, ReadOnlyBytecodeExecution, ReadOnlyCall, TimeInterval,
+use massa_models::api::{ AddressInfo, BlockInfo, BlockSummary, EndorsementInfo, EventFilter, NodeStatus,
+    OperationInfo, ReadOnlyExecution, TimeInterval,
 };
 use massa_models::clique::Clique;
 use massa_models::composite::PubkeySig;
@@ -34,6 +33,8 @@ use tracing::{info, warn};
 mod error;
 mod private;
 mod public;
+mod settings;
+pub use settings::APISettings;
 
 /// Public api component
 pub struct Public {
