+++ massa-node/src/settings.rs
@@ -3,10 +3,10 @@
 //! Build here the default node settings from the config file toml
 use std::path::PathBuf;
 
+use massa_api::APISettings;
 use massa_bootstrap::BootstrapSettings;
 use massa_consensus_exports::ConsensusSettings;
 use massa_models::{
-    api::APISettings,
     constants::{build_massa_settings, OPERATION_VALIDITY_PERIODS, THREAD_COUNT},
 };
 use massa_network_exports::NetworkSettings;
